# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Church ERP API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /discipleship/summary:
    get:
      operationId: getDiscipleshipSummary
      tags: [discipleship]
      summary: Discipleship 4x3 health matrix (active members only)
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscipleshipSummary"
  /discipleship/at-risk:
    get:
      operationId: getDiscipleshipAtRisk
      tags: [discipleship]
      summary: Members with red health in any area
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: query, name: area, required: false, schema: { $ref: "#/components/schemas/MemberArea" } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtRiskResponse"
  /discipleship/leaders:
    get:
      operationId: getDiscipleshipLeaders
      tags: [discipleship]
      summary: Members who are leaders in any area
      security: [{ cookieAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscipleshipLeadersResponse"
  /discipleship/by-leader/{leaderId}:
    get:
      operationId: getDiscipleshipByLeader
      tags: [discipleship]
      summary: Members under this leader (group by area)
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: leaderId, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AtRiskResponse"
  /discipleship/members/{id}/areas:
    get:
      operationId: getMemberAreas
      tags: [discipleship]
      summary: List 4 discipleship areas for a member
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberAreasResponse"
  /discipleship/members/{id}/areas/{area}:
    put:
      operationId: updateMemberArea
      tags: [discipleship]
      summary: Update one discipleship area for a member
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
        - { in: path, name: area, required: true, schema: { $ref: "#/components/schemas/MemberArea" } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberAreaRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberAreaEntry"
  /discipleship/members/{id}/history:
    get:
      operationId: getMemberAreaHistory
      tags: [discipleship]
      summary: Discipleship area transition history
      security: [{ cookieAuth: [] }]
      parameters:
        - { in: path, name: id, required: true, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberAreaHistoryResponse"
...
  /members/{id}:
    get:
      operationId: getMember
      tags: [members]
      summary: Get a single member
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberDetail"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateMember
      tags: [members]
      summary: Update a member
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMemberRequest"
      responses:
        "200":
          description: Member updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMember
      tags: [members]
      summary: Delete a member (Admin only)
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /members/{id}/history:
    get:
      operationId: getMemberHistory
      tags: [members]
      summary: Get member change history
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Member history entries
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MemberHistoryResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    MemberStatus:
      type: string
      enum: [ativo, disciplina, rol_apartado, falecido, demitido]
    MemberClassification:
      type: string
      enum: [comungante, nao_comungante]
    MemberReceptionMode:
      type: string
      enum:
        - profissao_fe
        - profissao_fe_batismo
        - carta_transferencia
        - jurisdicao_pedido
        - jurisdicao_ex_officio
        - restauracao
        - batismo_infantil
        - transferencia_menor
        - arrolamento_menor
    MemberMaritalStatus:
      type: string
      enum: [solteiro, casado, viuvo, divorciado, uniao_estavel]
    MemberExclusionReason:
      type: string
      enum:
        - transferencia
        - falecimento
        - exclusao_pedido
        - exclusao_disciplina
        - exclusao_abandono
        - ordenacao_ministerio
        - transferencia_responsaveis
        - profissao_fe_migracao
        - exclusao_abandono_responsaveis
...
    MemberLink:
      type: object
      properties:
        id: { type: string }
        fullName: { type: string }
        isExternal: { type: boolean }
        childMemberId: { type: string, nullable: true }
      required: [id, fullName]
    MemberGroupLink:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
      required: [id, name]
    MemberDetail:
      type: object
      properties:
        id: { type: string }
        fullName: { type: string }
        cpfMasked: { type: string }
        dateOfBirth: { type: string }
        sex: { type: string, enum: [masculino, feminino] }
        phone: { type: string }
        email: { type: string }
        addressZip: { type: string }
        addressStreet: { type: string }
        addressNumber: { type: string }
        addressComplement: { type: string }
        addressNeighborhood: { type: string }
        addressCity: { type: string }
        addressState: { type: string }
        # Eclesiástico
        classification:
          $ref: "#/components/schemas/MemberClassification"
        receptionMode:
          $ref: "#/components/schemas/MemberReceptionMode"
        receptionDate: { type: string }
        conversionYear: { type: integer }
        religiousOrigin: { type: string }
        infantBaptism: { type: boolean }
        infantBaptismChurch: { type: string }
        infantBaptismPastor: { type: string }
        parentsOrGuardians: { type: string }
        # Pessoal
        maritalStatus:
          $ref: "#/components/schemas/MemberMaritalStatus"
        spouseMemberId: { type: string }
        spouseName: { type: string }
        externalSpouseName: { type: string, nullable: true }
        academicEducation: { type: string }
        profession: { type: string }
        # Status / exclusão
        status:
          $ref: "#/components/schemas/MemberStatus"
        exclusionReason:
          $ref: "#/components/schemas/MemberExclusionReason"
        exclusionDate: { type: string }
        exclusionNotes: { type: string }
        exclusionLetterPath: { type: string }
        photoPath: { type: string }
        children:
          type: array
          items:
            $ref: "#/components/schemas/MemberLink"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/MemberGroupLink"
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - fullName
        - cpfMasked
        - classification
        - status
        - createdAt
        - updatedAt
...
    UpdateMemberRequest:
      type: object
      properties:
        fullName: { type: string }
        cpf: { type: string }
        dateOfBirth: { type: string }
        sex: { type: string, enum: [masculino, feminino] }
        phone: { type: string }
        email: { type: string }
        addressZip: { type: string }
        addressStreet: { type: string }
        addressNumber: { type: string }
        addressComplement: { type: string }
        addressNeighborhood: { type: string }
        addressCity: { type: string }
        addressState: { type: string }
        classification:
          $ref: "#/components/schemas/MemberClassification"
        receptionMode:
          $ref: "#/components/schemas/MemberReceptionMode"
        receptionDate: { type: string }
        conversionYear: { type: integer }
        religiousOrigin: { type: string }
        infantBaptism: { type: boolean }
        infantBaptismChurch: { type: string }
        infantBaptismPastor: { type: string }
        parentsOrGuardians: { type: string }
        maritalStatus:
          $ref: "#/components/schemas/MemberMaritalStatus"
        spouseMemberId: { type: string, nullable: true }
        externalSpouseName: { type: string, nullable: true }
        academicEducation: { type: string }
        profession: { type: string }
        status:
          $ref: "#/components/schemas/MemberStatus"
...
    MemberHistoryEntry:
      type: object
      properties:
        id:
          type: string
        memberId:
          type: string
        changedByUserId:
          type: string
        changeType:
          type: string
        fieldChanges:
          type: object
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - memberId
        - changedByUserId
        - changeType
        - createdAt
    MemberHistoryResponse:
      type: object
      properties:
        history:
          type: array
          items:
            $ref: "#/components/schemas/MemberHistoryEntry"
      required:
        - history
...
    MemberArea:
      type: string
      enum: [culto, pequeno_grupo, ministerio, ebd]
    MemberAreaHealth:
      type: string
      enum: [verde, amarelo, vermelho]
    DiscipleshipSummary:
      type: object
      properties:
        matrix:
          type: object
          additionalProperties:
            type: object
            additionalProperties: { type: integer }
      required: [matrix]
    AtRiskItem:
      type: object
      properties:
        memberId: { type: string }
        memberName: { type: string }
        area: { $ref: "#/components/schemas/MemberArea" }
        areaLabel: { type: string }
        healthStatus: { $ref: "#/components/schemas/MemberAreaHealth" }
        leaderMemberName: { type: string, nullable: true }
      required: [memberId, memberName, area, areaLabel, healthStatus]
    AtRiskResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/AtRiskItem" }
        total: { type: integer }
      required: [items, total]
    DiscipleshipLeaderItem:
      type: object
      properties:
        leaderMemberId: { type: string, nullable: true }
        leaderMemberName: { type: string, nullable: true }
        total: { type: integer }
      required: [total]
    DiscipleshipLeadersResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/DiscipleshipLeaderItem" }
      required: [items]
    MemberAreaEntry:
      type: object
      properties:
        id: { type: string }
        area: { $ref: "#/components/schemas/MemberArea" }
        areaLabel: { type: string }
        healthStatus: { $ref: "#/components/schemas/MemberAreaHealth" }
        leaderMemberId: { type: string, nullable: true }
        leaderMemberName: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        lastUpdatedAt: { type: string }
      required: [id, area, areaLabel, healthStatus]
    MemberAreasResponse:
      type: object
      properties:
        memberId: { type: string }
        memberName: { type: string }
        areas:
          type: array
          items: { $ref: "#/components/schemas/MemberAreaEntry" }
      required: [memberId, areas]
    UpdateMemberAreaRequest:
      type: object
      properties:
        healthStatus: { $ref: "#/components/schemas/MemberAreaHealth" }
        leaderMemberId: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        reason: { type: string, nullable: true }
    MemberAreaHistoryItem:
      type: object
      properties:
        id: { type: string }
        area: { $ref: "#/components/schemas/MemberArea" }
        areaLabel: { type: string }
        fromHealth:
          allOf: [{ $ref: "#/components/schemas/MemberAreaHealth" }]
          nullable: true
        toHealth: { $ref: "#/components/schemas/MemberAreaHealth" }
        reason: { type: string, nullable: true }
        createdAt: { type: string }
      required: [id, area, areaLabel, toHealth, createdAt]
    MemberAreaHistoryResponse:
      type: object
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/MemberAreaHistoryItem" }
      required: [items]

    # ─── MINISTRY GOAL SCHEMAS ─────────────────────────────────────────────
...